Rename the organization display and menu services
OrganizationDisplayer becomes OrgServices::Displayer, OrganizedServices
becomes OrgServices, and the two menu builders become
UserServices::MenuItemsOrg and UserServices::MenuItemsAccount, so the
pair sits together and sorts side by side.

Moves OrganizedHelper#retrieval_link_url and #registration_field_label
onto OrgServices::Displayer, and renames the menu translation scopes and
the org menu cache key to follow their classes.
