+++ app/components/emails/finished_registration/component.rb
@@ -162,7 +162,7 @@ module Emails
 
       def bike_url_with_token
         if @email_preview
-          OrganizedServices::EmailPreview::TOKEN_PATH
+          OrgServices::EmailPreview::TOKEN_PATH
         else
           bike_url(bike, t: @ownership.token, email:)
         end
@@ -170,7 +170,7 @@ module Emails
 
       def recovery_url
         if @email_preview
-          OrganizedServices::EmailPreview::TOKEN_PATH
+          OrgServices::EmailPreview::TOKEN_PATH
         else
           edit_bike_recovery_url(bike_id: bike.id, token: bike.fetch_current_stolen_record.find_or_create_recovery_link_token)
         end
