+++ app/components/emails/organization_invitation/component.rb
@@ -24,7 +24,7 @@ module Emails
       end
 
       def tokenized_url
-        @email_preview ? OrganizedServices::EmailPreview::TOKEN_PATH : new_user_url(email: invited_email)
+        @email_preview ? OrgServices::EmailPreview::TOKEN_PATH : new_user_url(email: invited_email)
       end
     end
   end
