+++ app/components/emails/parking_notification/component.rb
@@ -51,7 +51,7 @@ module Emails
       end
 
       def tokenized_url
-        @email_preview ? OrganizedServices::EmailPreview::TOKEN_PATH : helpers.retrieval_link_url(@parking_notification)
+        @email_preview ? OrgServices::EmailPreview::TOKEN_PATH : OrgServices::Displayer.retrieval_link_url(@parking_notification)
       end
 
       def map_url
