+++ app/components/emails/partial_registration/component.rb
@@ -24,7 +24,7 @@ module Emails
       end
 
       def tokenized_url
-        @email_preview ? OrganizedServices::EmailPreview::TOKEN_PATH : new_bike_url(b_param_token: @b_param.id_token)
+        @email_preview ? OrgServices::EmailPreview::TOKEN_PATH : new_bike_url(b_param_token: @b_param.id_token)
       end
 
       def organization_snippet_body
