+++ app/components/register/step1/component.rb
@@ -60,13 +60,13 @@ module Register
 
       # owner_email is the setting bikes/new labels its email field with
       def email_label
-        helpers.registration_field_label(organization, "owner_email", strip_tags: true) ||
+        OrgServices::Displayer.registration_field_label(organization, "owner_email", strip_tags: true) ||
           (translation(".email_school", org_name: organization.short_name) if organization&.school?) ||
           translation(".email")
       end
 
       def email_placeholder
-        helpers.registration_field_label(organization, "email_placeholder", strip_tags: true) ||
+        OrgServices::Displayer.registration_field_label(organization, "email_placeholder", strip_tags: true) ||
           translation(".email_placeholder")
       end
