+++ app/helpers/organized_helper.rb
@@ -108,23 +108,4 @@ module OrganizedHelper
 
     l datetime, format: :dotted
   end
-
-  def retrieval_link_url(obj)
-    if obj.is_a?(ParkingNotification)
-      return nil if obj.retrieval_link_token.blank?
-
-      bike_url(obj.bike.to_param, parking_notification_retrieved: obj.retrieval_link_token)
-    elsif obj.is_a?(GraduatedNotification)
-      return nil if obj.marked_remaining_link_token.blank?
-
-      bike_url(obj.bike.to_param, graduated_notification_remaining: obj.marked_remaining_link_token)
-    end
-  end
-
-  def registration_field_label(organization = nil, field_slug = nil, strip_tags: false)
-    txt = organization&.registration_field_labels&.dig(field_slug.to_s)
-    return nil unless txt.present?
-
-    strip_tags ? Binxtils::InputNormalizer.sanitize(txt) : txt.html_safe
-  end
 end
