+++ app/services/org_services/email_preview.rb
@@ -1,4 +1,4 @@
-module OrganizedServices
+module OrgServices
   module EmailPreview
     extend Functionable
