+++ app/views/registrations/embed.html.haml
@@ -1,6 +1,6 @@
 = form_for @b_param, { url: registrations_path, action: "create" } do |f|
   .form-group
-    - email_placeholder = registration_field_label(current_organization, "owner_email", strip_tags: true)
+    - email_placeholder = OrgServices::Displayer.registration_field_label(current_organization, "owner_email", strip_tags: true)
     - email_placeholder ||= t(".owner_email", owner: current_organization&.school? ? current_organization.short_name : t(".your"))
     = f.email_field :owner_email, required: true, value: @owner_email, placeholder: email_placeholder, class: "form-control"
