+++ app/views/shared/_claim_message.html.erb
@@ -3,7 +3,7 @@
 <% skip_to_see ||= false %>
 
 <p>
-  <% if OrganizationDisplayer.avatar?(ownership.organization) %>
+  <% if OrgServices::Displayer.avatar?(ownership.organization) %>
     <%= image_tag ownership.organization.avatar.url(:medium), alt: ownership.organization.name, class: "claim-message-org-avatar" %>
   <% end %>
