+++ config/locales/en.yml
@@ -4790,9 +4790,6 @@ en:
       passwordless_help: We'll email you a secure link to sign in.
       sign_in_without_password: Sign in without a password
   shared:
-    account_menu_items:
-      marketplace_messages: Marketplace messages
-      switch_to_org: Switch to %{org_name} admin
     bike_search_form:
       anywhere: Anywhere
       doesnt_look_like_serial: >-
@@ -4972,7 +4969,10 @@ en:
       you: You
       your_bikes: Your bikes
       your_preferences: Your preferences
-    organized_menu_items:
+    menu_items_account:
+      marketplace_messages: Marketplace messages
+      switch_to_org: Switch to %{org_name} admin
+    menu_items_org:
       add_a_bike: Add a bike
       add_an_impounded_vehicle: Add an Impounded Vehicle
       ascend_imports: Ascend Imports
