+++ config/locales/translation.nb.yml
@@ -4968,7 +4968,7 @@ nb:
       you: Du
       your_bikes: Syklene dine
       your_preferences: Dine preferanser
-    organized_menu_items:
+    menu_items_org:
       add_a_bike: Legg til en sykkel
       ascend_imports: Ascend Import
       bulk_imports: Bulk import
@@ -5019,7 +5019,7 @@ nb:
       manufacturer_required: Produsenten er pålagt å registrere seg
       name_required: Eiernavn kreves for registrering
       unable_to_save: Vi kan ikke lagre registreringen. Prøv på nytt.
-    account_menu_items:
+    menu_items_account:
       marketplace_messages: Markedsplassmeldinger
       switch_to_org: Bytt til %{org_name}-administrator
   stolen:
